requirements.txt
selenium==4.15.2
webdriver-manager==4.0.1
pytest==7.4.3
pytest-mock==3.12.0
python-dotenv==1.0.0
pydantic==2.5.0
